Skip generating lease file if it already exists
Because the DHCP agent does not know the IAID (identity association
identifier) of assigned IPv6 addresses it's not possible to generate the
lease file including IPv6 leases. Because of this IPv6 addresses are excluded
when generating the lease file in case of DHCP agent restarts. This causes
DHCPv6 clients to fail to RENEW their lease and to go through a full address
discovery cycle with possible short connectivity disruption.

Generating the lease file is not neccessary if a lease file already
exists as is the case if the agent was just restarted due to a
configuration change or host reboot. This commit skips the lease file
generation if it already exists.

A lease file is still generated if an agent is migrated to a different
host.

This commit complements the fix implemented in
Ib1b2f284ab81f1c4af7b08b5257b45a3f6e79c3e which just skips the IPv6
leases as otherwise the lease file would be invalid and all leases would
be lost. It does not change the behavior for still valid IPv4 leases
as they are present in the lease file.

Closes-Bug: #1722126
Related-Change: Ib1b2f284ab81f1c4af7b08b5257b45a3f6e79c3e

Change-Id: I40761b30563749251b9d74731bbe7a80a124da89
(cherry picked from commit 0ddbb5386ddae0abb343871a7e82a25b6d7d8505)