Faster VLAN children lookup for linuxbridge agent
Before this commit, the `Plumber._get_vlan_children` function in the
linuxbridge agent loaded all the network interfaces to look for
parent/child relationships.

This takes seconds on a system with hundreds of network devices and
causes the wiring the linuxbridge agent performs, to be a lot slower
than it needs to be.

This change passes filters down to `pyroute2` to avoid loading so many
interfaces. Instead, only the parent is loaded once, and then the
`IFLA_LINK` attribute is used (via the `link` argument), to look for
interfaces that point to this parent.