Use transient systemd units in Process fixture
This makes the damon processes spawned by fullstack tests more similar
to those used in a real deployment. It creates a systemd unit for each
daemon on the fly and combines all daemons of the same host in a systemd
slice. The units are automatically removed on service shutdown.

This also better ensures that there are no left over processes after
failed test cases as all processes started in a systemd unit are cleaned
up when the unit is stopped.

This change is a prerequisite to implement methods to kill or shutdown a
host. Without this change it's impossible to find out which processes
belong to a agent and to kill all these processes. This is important to
eg. kill keepalived processes started by L3 agents.

The timeout for the L2 agent restart has to be extended slightly because
the service is no longer killed with SIGKILL but properly shut down.

Change-Id: Ieb21de8a370aca24abc216f55dbd98e553a65d05
(cherry picked from commit 4a02848228a4dce926fc2c7c5271bcbf0bd7c3a4)